4-RS-Algorithms/requirements.txt
numpy
configargparse
pandas
spacy==3.1.0
scikit-learn
lenskit
cffi
implicit==0.4.8
lightfm==1.17
rdflib==6.1.1
ssmpy==0.2.5
mysql-connector-python==8.0.32
pymysql
sqlalchemy